requirements.txt
streamlit
requests
pandas
Pillow
pymongo
python-dotenv
protobuf
